data-ingestion-agents/data-extraction-agent/requirements.txt
# ==============================================================================
# Requirements for data extraction agent
# ==============================================================================
# Prerequisites:
# ------------------------------------------------------------------------------
# Install pip if not already installed:
# sudo apt install python3-pip
#
# Create a virtual environment (for package isolation from system-wide packages):
# Install venv module if not already installed:
#
# sudo apt install python3.12-venv
# python3 -m venv .venv
# source .venv/bin/activate


# ==============================================================================
# Required packages: make install
# ==============================================================================
# Azure Cosmos DB SDK
azure-core==1.36.0
azure-cosmos==4.14.0
azure-identity==1.25.1

# Common packages
pydantic-settings==2.11.0
urllib3==2.5.0